Add files via upload
(1) Add the name of the pdf file. The file path entry has a comment for indication.  (cell #10)
(2) The notebook was used on Kaggle with P100 GPU
(3) test_rag(qa, query): is used to ask the LLM a question about something in your pdf, qa->RetrievalQA chain, query->string containing your question
(4) query_database(query): is used to perform a similarity search on your vectordb and return the top-4 results